refactor: Customer Column 추가
