--- src/main/resources/application-test.yml
@@ -1,14 +1,0 @@
-spring:
-  h2:
-    console:
-      enabled: true
-  jpa:
-    database: h2
-    generate-ddl: true
-    hibernate:
-      ddl-auto: create-drop
-  datasource:
-    driver-class-name: org.h2.Driver
-    url: jdbc:h2:mem:testdb;MODE=MySQL;
-    username: sa
-    password:
