+++ src/test/java/com/kdt/weeklyjpa/domain/CustomerRepositoryTest.java
@@ -1,20 +1,19 @@
 package com.kdt.weeklyjpa.domain;
 
+import com.kdt.weeklyjpa.domain.customer.Customer;
+import com.kdt.weeklyjpa.domain.customer.CustomerRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.transaction.annotation.Transactional;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
 import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-@SpringBootTest
-@ActiveProfiles("test")
+@DataJpaTest
 class CustomerRepositoryTest {
 
     @Autowired
@@ -26,15 +25,14 @@ class CustomerRepositoryTest {
     }
 
     @Test
-    @Transactional
     @DisplayName("Customer_생성_테스트")
     void saveTest() {
         // given
-        Customer customer = new Customer("Youngmyung", "Kim");
+        Customer customer = new Customer("Youngmyung", "Kim", "010-1234-5678");
 
         // when
         Customer saved = customerRepository.save(customer);
-        Optional<Customer> foundById = customerRepository.findById(saved.getId());
+        Optional<Customer> foundById = customerRepository.findById(saved.getCustomer_id());
 
         // then
         assertThat(foundById)
@@ -48,9 +46,9 @@ class CustomerRepositoryTest {
     @DisplayName("Customers_조회_테스트")
     void getCustomerTest() {
         // given
-        Customer customer1 = new Customer("Youngmyung", "Kim");
-        Customer customer2 = new Customer("Test1", "Hong");
-        Customer customer3 = new Customer("Test2", "Kwon");
+        Customer customer1 = new Customer("Youngmyung", "Kim", "010-1234-5678");
+        Customer customer2 = new Customer("Test1", "Hong", "010-1111-2222");
+        Customer customer3 = new Customer("Test2", "Kwon", "010-3333-4444");
 
 
         // when
@@ -64,15 +62,14 @@ class CustomerRepositoryTest {
     }
 
     @Test
-    @Transactional
     @DisplayName("Customer_update_테스트")
     void updateCustomerTest() {
         // given
-        Customer customer = new Customer("Youngmyung", "Kim");
+        Customer customer = new Customer("Youngmyung", "Kim", "010-1234-5678");
 
         // when
         customerRepository.save(customer);
-        Customer savedCustomer = customerRepository.findById(customer.getId()).get();
+        Customer savedCustomer = customerRepository.findById(customer.getCustomer_id()).get();
         savedCustomer.updateFirstName("NewFirstName");
 
         // then
@@ -83,13 +80,13 @@ class CustomerRepositoryTest {
     @DisplayName("Customer_delete_테스트")
     void deleteCustomerTest() {
         // given
-        Customer customer1 = new Customer("Youngmyung", "Kim");
-        Customer customer2 = new Customer("Test1", "Hong");
+        Customer customer1 = new Customer("Youngmyung", "Kim", "010-1234-5678");
+        Customer customer2 = new Customer("Test1", "Hong", "010-7777-9999");
 
         // when
         customerRepository.save(customer1);
         customerRepository.save(customer2);
-        customerRepository.deleteById(customer1.getId());
+        customerRepository.deleteById(customer1.getCustomer_id());
         List<Customer> customers = customerRepository.findAll();
 
         // then
